+++ src/Scripts/RoomDataTartarus.lua
@@ -807,7 +807,7 @@ local roomModifications = {
 
 		LegalEncounters = {
 			-- TODO: Enable after testing
-			-- "Story_Sisyphus_01",
+			"Story_Sisyphus_01",
 			-- "Story_Sisyphus_01",
 			-- "Story_Sisyphus_01",
 			-- "Story_Sisyphus_01",